Refactor parsing boombox opts
